登录、缓存数据、加载上一次的数据

+++ src/user/user.reducer.ts
@@ -1,3 +1,11 @@
+import { initProjectState } from './../project/project.actions';
+import { initCompanyState } from './../company/company.actions';
+import { Company } from './../company/company.modal';
+import { UserState } from './user.reducer';
+import { MroUtils } from './../common/mro-util';
+import { INIT_APP_STORE, InitAppStoreAction } from './../app/app.actions';
+import { CompanyState, CompanyReducer } from './../company/company.reducer';
+import { ProjectState, ProjectReducer } from './../project/project.reducer';
 import { Project } from './../project/project.modal';
 import { tableNames } from './../providers/db-operation/mro.tables';
 import { EpicDependencies, AppState } from './../app/app.reducer';
@@ -6,7 +14,7 @@ import { RequestOptions } from '@angular/http';
 import { MroError, MroErrorCode, generateMroError } from './../app/mro-error-handler';
 import { Http } from '@angular/http';
 import { User } from "./user.modal";
-import { Action, Store } from "redux";
+import { Action, Store, combineReducers } from "redux";
 import * as UserActions from "./user.actions";
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/switchMap';
@@ -15,6 +23,7 @@ import { ActionsObservable, combineEpics, Epic } from "redux-observable";
 import * as Apis from "../providers/api/api";
 import { inspect } from "util";
 import 'rxjs/add/observable/of';
+import 'rxjs/add/observable/from';
 import 'rxjs/add/operator/do';
 import 'rxjs/add/operator/mapTo';
 import 'rxjs/add/operator/catch';
@@ -22,34 +31,59 @@ import 'rxjs/add/operator/delay';
 import * as ProjectActions from "../project/project.actions";
 export interface UserState {
   currentUser: User;
-  isTokenValid: boolean;
-  isFetching: boolean;
+  userProject: ProjectState
+  userCompany: CompanyState
 }
-export const UserReducer = function (state: UserState = null, action: Action): UserState {
+export function UserReducer(user: User = null, action: Action): User {
   switch (action.type) {
     case UserActions.SET_CURRENT_USER: {
-      const user = (<UserActions.SetCurrentUserAction>action).user;
-      return {
-        ...state,
-        currentUser: user
+       user=(<UserActions.SetCurrentUserAction>action).user;
+      return user;
+    }
+    case UserActions.INIT_USER_STATE: {
+      let user2=(<InitAppStoreAction>action).userSate.currentUser;
+      console.log("UserActions.INIT_USER_STATE",user2);
+      if(user2){
+        return user2;
       }
+      return user;
     }
     default: {
-      return state
+      return user;
     }
   }
 }
+export const initUserReducerEpic = (action$: ActionsObservable<Action>, store: Store<AppState>, deps: EpicDependencies) => {
+  return action$.ofType(UserActions.INIT_USER_STATE)
+    .switchMap((action: InitAppStoreAction) => {
+      console.log("初始化，initUserReducerEpic");
+      return Observable.from([
+        initCompanyState(action.userSate.userCompany),
+        initProjectState(action.userSate.userProject)
+      ]);
+    })
+}
+export const RootUserReducer = combineReducers({
+  currentUser: UserReducer,
+  userProject: ProjectReducer,
+  userCompany: CompanyReducer,
+})
 const userLoginEpic: Epic<Action, AppState, EpicDependencies> = (action$: ActionsObservable<Action>, store, deps: EpicDependencies) => {
   return action$.ofType(UserActions.LOGIN_USER)
     .switchMap((action: UserActions.UserLoginAction) => {
       inspect({ ...action.payload })
       const loading = deps.loadingCtrl.create({
-        content:'正在登陆...'
+        content: '正在登陆...',
+        dismissOnPageChange: true
       });
       loading.present();
       return deps.http.post(Apis.Api_login, action.payload)
-        .do(()=>loading.dismiss())
-        .map(res => UserActions.setCurrentUser(res.json().data as User))
+        .do(() => loading.dismiss())
+        .map(res => {
+          const user= res.json().data as User;
+          MroUtils.setLoginUserId(user.id+"");
+          return UserActions.setCurrentUser(user);
+        })
         .catch((e: Error) => {
           console.error(e);
           loading.dismiss();
@@ -62,24 +96,24 @@ const userLoginEpic: Epic<Action, AppState, EpicDependencies> = (action$: Action
 }
 const setUserInfoEpic: Epic<Action, AppState, EpicDependencies> = (action$: ActionsObservable<Action>, store: Store<AppState>, deps: EpicDependencies) => {
   return action$.ofType(UserActions.SET_CURRENT_USER)
-    .switchMap((action) => {
-      console.log("set User")
-      const sqls: any[] = [`delete from ${tableNames.eam_user}`];
-      const user = store.getState().currentUser;
+    .switchMap((action: UserActions.SetCurrentUserAction) => {
+      console.log("set User", action.user)
+      const sqls: any[] = [[`delete from ${tableNames.eam_user} where  userId=?`, [action.user.id]]];
+      const user = action.user;
       sqls.push([`insert into ${tableNames.eam_user}(
         userJson,
         lastLoginTime,
         userId
       )values(?,?,?)`,
-      [JSON.stringify(user.currentUser), Date.now(), user.currentUser.id]
-      ])
+      [JSON.stringify(user), Date.now(), user.id]])
       const loading = deps.loadingCtrl.create({
-        content:'正在写入用户信息到数据库...'
+        content: '正在写入用户信息到数据库...',
+        dismissOnPageChange: true
       });
       loading.present();
       return deps.dbOperation
         .sqlBatch(sqls)
-        .do(()=>loading.dismiss())
+        .do(() => loading.dismiss())
         .mapTo(ProjectActions.fetchProjects())
         .catch((e: Error) => {
           console.error(e);
@@ -89,4 +123,4 @@ const setUserInfoEpic: Epic<Action, AppState, EpicDependencies> = (action$: Acti
         })
     })
 }
-export const rootUserEpic = combineEpics(...[userLoginEpic, setUserInfoEpic]);
+export const rootUserEpic = combineEpics(...[userLoginEpic, setUserInfoEpic, initUserReducerEpic]);
